requirements.txt
requests
pandas
plotly
plotnine
python-dotenv
folium
ipython
jupyter
ipykernel